requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
starlette~=0.47.3
pydantic==2.11.7
pandas~=2.3.2
numpy==2.3.2
opencc==1.1.9
pypinyin~=0.55.0

SQLAlchemy~=2.0.43
future~=1.0.0
openpyxl~=3.1.5
xlrd~=2.0.1
h11~=0.16.0
pip~=24.3.1
typing_extensions~=4.14.1
attrs~=24.2.0
retry~=0.9.2
cryptography~=45.0.3
filelock~=3.18.0
pytz~=2025.2
sniffio~=1.3.1
trio~=0.26.2
outcome~=1.3.0.post0
setuptools~=70.0.0
cffi~=1.17.1
pyxlsb~=1.0.10
lxml~=5.2.2
MarkupSafe~=3.0.2
python-dateutil~=2.9.0.post0
py~=1.11.0
pydantic_core~=2.33.2
requests~=2.32.3
websockets~=15.0.1
anyio~=4.9.0
click~=8.2.1
httptools~=0.6.4
wsproto~=1.2.0
watchfiles~=1.1.0
altgraph~=0.17.4
six~=1.17.0
packaging~=25.0
itsdangerous~=2.2.0
PyYAML~=6.0.1
Jinja2~=3.1.6
httpx~=0.28.1
pefile~=2023.2.7
wheel~=0.43.0
platformdirs~=4.3.8
Pygments~=2.19.2
hanziconv~=0.3.2
passlib~=1.7.4
email_validator~=2.2.0
python-jose~=3.5.0
gunicorn==21.2.0
python-multipart
redis~=7.1.0

APScheduler~=3.11.2
python-docx==1.2.0
bcrypt==3.2.0

# Praat acoustic analysis dependencies
praat-parselmouth>=0.4.3
scipy>=1.10.0

# VillagesML machine learning dependencies
scikit-learn>=1.3.0
networkx>=3.0

# Performance optimization for string matching
rapidfuzz>=3.0.0

# Analytics dependencies
geoip2>=4.7.0
user-agents>=2.2.0